// src/models/redux/enums.ts
export enum UserActionEnum {
  SET_LOADING = 'SET_LOADING',
  USER_REGISTRATE = 'USER_REGISTRATE',
  USER_LOGIN = 'USER_LOGIN',
  USER_LOGOUT = 'USER_LOGOUT',
  USER_FORGOT = 'USER_FORGOT',
  USER_RESTORE = 'USER_RESTORE',
  USER_EDIT = 'USER_EDIT',
  USER_ERROR_OPEN_SNACKBAR = 'USER_ERROR_OPEN_SNACKBAR',
  USER_AUTHORIZED_ERROR = 'USER_AUTHORIZED_ERROR',
  USER_AUTHORIZED_CHECK = 'USER_AUTHORIZED_CHECK',
  CHANGE_AMOUNT_OF_CART_PRODUCT = 'CHANGE_AMOUNT_OF_CART_PRODUCT',
  ADD_OR_REMOVE_PRODUCT_FROM_SHOPPING_CART = 'ADD_OR_REMOVE_PRODUCT_FROM_SHOPPING_CART',
  CLEAR_SHOPPING_CART = 'CLEAR_SHOPPING_CART',
  ADD_OR_REMOVE_PRODUCT_FROM_FAVORITE = 'ADD_OR_REMOVE_PRODUCT_FROM_FAVORITE',
  GET_FAVORITE_FROM_LOCAL_STORAGE = 'GET_FAVORITE_FROM_LOCAL_STORAGE',
  SAVE_FAVORITE_IN_LOCAL_STORAGE = 'SAVE_FAVORITE_IN_LOCAL_STORAGE',
  DELETE_PRODUCT_FROM_STORAGE = 'DELETE_PRODUCT_FROM_STORAGE',
  CLEAR_FAVORITE_LOCAL_STORAGE = 'CLEAR_FAVORITE_LOCAL_STORAGE',
  CLEAR_FAVORITE = 'CLEAR_FAVORITE',
}

export enum CategoryActionEnum {
  CATEGORIES_LOADING = 'CATEGORIES_LOADING',
  GET_ALL_CATEGORIES = 'GET_ALL_CATEGORIES',
  GET_PART_OF_CATEGORIES = 'GET_PART_OF_CATEGORIES',
  CATEGORIES_ERRORS = 'CATEGORIES_ERRORS',
  CATEGORIES_ERROR_CLOSE_SNACKBAR = 'USER_ERROR_CLOSE_SNACKBAR',
}

export enum ProductsActionEnum {
  PRODUCTS_LOADING = 'PRODUCTS_LOADING',
  GET_ALL_PRODUCTS = 'GET_ALL_PRODUCT',
  GET_CART = 'GET_CART',
  ADD_TO_SHOPPING_CART = 'ADD_TO_SHOPPING_CART',
  GET_PART_OF_PRODUCTS = 'GET_PART_OF_PRODUCTS',
  FIND_PRODUCT = 'FIND_PRODUCT',
  PRODUCTS_ERRORS = 'PRODUCTS_ERRORS',
  GET_FAVORITES = 'GET_FAVORITES',
  PRODUCTS_ERROR_CLOSE_SNACKBAR = 'PRODUCTS_ERROR_CLOSE_SNACKBAR',
  GET_PRODUCTS_BY_ARRAY_ID = 'GET_PRODUCTS_BY_ARRAY_ID',
}